test(admin): consolidate the e2e suite by section
Reorganize 22 admin specs into 10 section-aligned suites so each
surface has one home: auth (login, bootstrap, device grant, email-change
feedback), users (list, invite, edit, email change, API tokens),
taxonomies, settings, dashboard, editor (chrome, slash inserts,
accessibility, autosave/publish contracts), editor-actions (authoring +
drafts of published entries), editor-collab (stale-draft dialog,
revision preview, co-author indicator), plain-form (non-editor types +
capability gating), and plugins (catch-all route, runtime alias seam,
block registration).

Delete low-value coverage: the per-block slash-insert specs collapse
into one parameterized loop, the desktop chrome visibility checklist and
duplicate autosave-envelope test go away, the weaker dispatchEvent
detach twin merges into the real-click variant, and the six
screenshot + waitForTimeout pairs (tmp/*.png, never asserted) are
removed.

Conventions: drop the stale v2/spike describe naming and issue refs in
titles, replace the remaining getByRole/text= selectors with testids
(adding users-list-row-you to the You badge), hoist shared fixtures
(insertViaSlash, publishedEntry, publishedEntryRpcBody, rpcConflictBody)
into e2e/support, and collapse the hand-rolled rpc route ladders into a
mockRpc floor with targeted page.route overrides for error and stateful
endpoints. The back-button test now performs a real click instead of
asserting the href attribute.
